+++ Packages/com.madratzz.scriptableobject.event.variables/Tests/EditMode/EventVariableTests.cs
@@ -108,5 +108,49 @@ namespace Madratzz.Tests.EventVariables
             Assert.DoesNotThrow(() => variable.RemoveListener(() => { }));
             Object.DestroyImmediate(variable);
         }
+
+        [Test]
+        public void IntWithEvent_SetValue_RaisesValueChanged()
+        {
+            var variable = ScriptableObject.CreateInstance<IntWithEvent>();
+            Wire(variable);
+            int calls = 0;
+            _valueChanged.Handler += () => calls++;
+
+            variable.SetValue(7);
+
+            Assert.AreEqual(1, calls);
+            Assert.AreEqual(7, variable.GetValue());
+            Object.DestroyImmediate(variable);
+        }
+
+        [Test]
+        public void IntWithEvent_ApplyChange_RaisesValueChanged()
+        {
+            // Incrementing is the common path for a score; without the
+            // ApplyChange override the value would move without notifying.
+            var variable = ScriptableObject.CreateInstance<IntWithEvent>();
+            Wire(variable);
+            variable.SetValue(10);
+            int calls = 0;
+            _valueChanged.Handler += () => calls++;
+
+            variable.ApplyChange(5);
+
+            Assert.AreEqual(1, calls);
+            Assert.AreEqual(15, variable.GetValue());
+            Object.DestroyImmediate(variable);
+        }
+
+        [Test]
+        public void IntWithEvent_WithNoEventAssigned_DoesNotThrow()
+        {
+            var variable = ScriptableObject.CreateInstance<IntWithEvent>();
+
+            Assert.DoesNotThrow(() => variable.SetValue(3));
+            Assert.DoesNotThrow(() => variable.ApplyChange(1));
+            Assert.AreEqual(4, variable.GetValue());
+            Object.DestroyImmediate(variable);
+        }
     }
 }